% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_pdf_table.R
\name{extract_pdf_table}
\alias{extract_pdf_table}
\title{Read data table from a pdf}
\usage{
extract_pdf_table(file_path, word_start, word_end, names_from, values_from)
}
\arguments{
\item{file_path}{path to pdf}

\item{word_start}{word that marks the start of the table (first word in the table of interest)}

\item{word_end}{word that marks the end of the table (first word after the table ends)}

\item{names_from}{table columns that are used to construct names for values
(eg. "CD4+", "Bead events", etc...)}

\item{values_from}{table columns that contain actual values (eg. "Percent", "Value/AbsCnt").
Currently all values are assumed to be numeric.}
}
\value{
a tibble with extracted data values, and columns 'patient', 'sample' and 'file'
}
\description{
Read data table from a pdf
}
\examples{
# Get a list of all PDF files in a folder (in this case our testdata folder)
folder <- system.file("testdata", "pdfs", package = "pdfdata")
files <- list.files(folder, pattern = "\\\\.pdf", full.names = TRUE)

pdf_tables <- purrr::map(files,
extract_pdf_table,
word_start = "Parameter",
word_end = "QC Messages",
names_from = "Parameter",
values_from = c("Percent", "Value/AbsCnt")) \%>\%
  # Combine the rows from each file
  dplyr::bind_rows()

pdf_tables
}
