% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_pdf_table.R
\name{facs_to_table}
\alias{facs_to_table}
\title{Function that takes one filepath, reads the text, and converts to a table}
\usage{
facs_to_table(file)
}
\arguments{
\item{file}{path to pdf file}
}
\value{
tibble with data values from table
}
\description{
Function that takes one filepath, reads the text, and converts to a table
}
\examples{
\dontrun{
facs_to_table(file_path)
}
}
