% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_pdf_table.R
\name{extract_pdf_table}
\alias{extract_pdf_table}
\title{Read data table from a pdf}
\usage{
extract_pdf_table(file_path)
}
\arguments{
\item{file_path}{path to pdf}
}
\value{
a tibble with extracted data values, and columns 'patient', 'sample' and 'file'
}
\description{
Read data table from a pdf
}
\examples{
# Get a list of all PDF files in a folder (in this case our testdata folder)
folder <- system.file("testdata", "pdfs", package = "pdfdata")
files <- list.files(folder, pattern = "\\\\.pdf", full.names = TRUE)

pdf_tables <- purrr::map(files, extract_pdf_table) \%>\%
  # Combine the rows from each file
  dplyr::bind_rows()

pdf_tables
}
